Änderungen im Report Generator, hinzufügen der Klassen AllContactsInSystemReport und AllContactsOfUserReport
