+++ ITProject_Group11_Jabics/src/de/hdm/group11/jabics/shared/report/AllContactsInSystemReport.java
@@ -1,5 +1,25 @@
 package de.hdm.group11.jabics.shared.report;
 
-public class AllContactsInSystemReport {
+import java.util.ArrayList;
 
+public class AllContactsInSystemReport {
+	
+	private ArrayList<AllContactsOfUserReport> subReports;
+	
+	public void addReport(AllContactsOfUserReport cr) {
+		subReports.add(cr);
+	}
+	
+	public void removeReport(AllContactsOfUserReport cr) {
+		subReports.remove(cr);
+	}
+	
+	public ArrayList<AllContactsOfUserReport> getSubReports() {
+		return subReports;
+	}
+
+	public void setSubReports(ArrayList<AllContactsOfUserReport> subReports) {
+		this.subReports = subReports;
+	}
+	
 }
